Develop - update

+++ CSH/library/api/fermat-csh-api/src/main/java/com/bitdubai/fermat_csh_api/layer/csh_cash_money_transaction/unhold/interfaces/CashUnholdTransactionParameters.java
@@ -1,7 +1,8 @@
 package com.bitdubai.fermat_csh_api.layer.csh_cash_money_transaction.unhold.interfaces;
 
-import com.bitdubai.fermat_api.layer.world.enums.FiatCurrency;
+import com.bitdubai.fermat_api.layer.all_definition.enums.FiatCurrency;
 
+import java.math.BigDecimal;
 import java.util.UUID;
 
 /**
@@ -37,7 +38,7 @@ public interface CashUnholdTransactionParameters {
      * Returns the amount of the transaction
      * @return      Amount of the transaction
      */
-    float getAmount();
+    BigDecimal getAmount();
 
     /**
      * Returns the currency of the transaction (e.g. USD, EUR)
